refactor: refactor metrics modules
