+++ packages/kit/src/views/Developer/pages/TabDeveloper.tsx
@@ -12,7 +12,7 @@ import {
   YStack,
 } from '@onekeyhq/components';
 import type { IPageNavigationProp } from '@onekeyhq/components/src/layouts/Navigation';
-import { getMeasureTime } from '@onekeyhq/shared/src/modules3rdParty/react-native-metrix';
+import { useMeasureTime } from '@onekeyhq/shared/src/modules3rdParty/metrics';
 import platformEnv from '@onekeyhq/shared/src/platformEnv';
 import { EAppSettingKey } from '@onekeyhq/shared/src/storage/appSetting';
 import appStorage from '@onekeyhq/shared/src/storage/appStorage';
@@ -61,7 +61,7 @@ function PartContainer({
 }
 
 function StartTimePanel() {
-  const { jsBundleLoadedTime, fpTime } = getMeasureTime();
+  const { jsBundleLoadedTime, fpTime } = useMeasureTime();
   return (
     <PartContainer title="Startup Time(ms)">
       <Text>JS Loaded Time: {jsBundleLoadedTime}</Text>
