+++ packages/shared/src/modules3rdParty/metrics/metrix.native.ts
@@ -1,0 +1,8 @@
+// Replace by performance.reactNativeStartupTiming.initializeRuntimeStart;
+//  in React Native 0.73
+export {
+  getTimeSinceStartup,
+  onUpdate,
+  start,
+  stop,
+} from 'react-native-metrix';
