--- packages/shared/src/modules3rdParty/react-native-metrix/metrix.native.ts
@@ -1,6 +1,0 @@
-export {
-  getTimeSinceStartup,
-  onUpdate,
-  start,
-  stop,
-} from 'react-native-metrix';
